backend/requirements.txt
# Enthält alle benötigten Python-Abhängigkeiten für das Backend (FastAPI, OpenAI, etc.)
fastapi
uvicorn[standard]
python-dotenv==1.0.0
openai
slowapi
pydantic
sqlalchemy
aiosqlite
pymysql==1.1.0
